requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
streamlit==1.39.0
yfinance==0.2.44
newspaper3k==0.2.8
langchain==0.3.1
langgraph==0.2.14
faiss-cpu==1.8.0.post1
transformers==4.44.2
speechrecognition==3.10.4
gtts==2.5.3
pandas==2.2.3
requests==2.32.3
numpy==1.26.4